Restructure package for distributed backends
